Moved the gameFactory module into game module
